requirements.txt
accelerate==1.8.1
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.12.14
aiosignal==1.4.0
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.9.0
argcomplete==3.6.2
async-timeout==5.0.1
attrs==25.3.0
bcrypt==4.3.0
boto3==1.39.0
botocore==1.39.0
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cryptography==45.0.4
cycler==0.12.1
dataclasses-json==0.6.7
Deprecated==1.2.18
diffusers==0.34.0
docstring_parser==0.16
einops==0.8.1
evo==1.31.1
exceptiongroup==1.3.0
executing==2.2.0
fastapi==0.115.14
ffmpy==0.6.0
filelock==3.13.1
flash_attn==2.8.0.post2
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2024.6.1
gitdb==4.0.12
GitPython==3.1.44
gradio==5.35.0
gradio_client==1.10.4
groovy==0.1.2
h11==0.16.0
hf-xet==1.1.5
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.33.1
hydra-core==1.3.2
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
importlib_metadata==8.7.0
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.5.1
jsonschema==4.24.0
jsonschema-specifications==2025.4.1
kiui==0.2.16
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.2
lightning-utilities==0.14.3
llvmlite==0.44.0
lz4==4.4.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.26.1
matplotlib==3.10.3
mdurl==0.1.2
megfile==4.1.8
mpmath==1.3.0
multidict==6.6.3
mypy_extensions==1.1.0
natsort==8.4.0
networkx==3.3
ninja==1.11.1.4
numba==0.61.2
numexpr==2.11.0
numpy==2.1.2
nvidia-cublas-cu11==11.11.3.6
nvidia-cuda-cupti-cu11==11.8.87
nvidia-cuda-nvrtc-cu11==11.8.89
nvidia-cuda-runtime-cu11==11.8.89
nvidia-cudnn-cu11==9.1.0.70
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.3.0.86
nvidia-cusolver-cu11==11.4.1.48
nvidia-cusparse-cu11==11.7.5.86
nvidia-nccl-cu11==2.21.5
nvidia-nvtx-cu11==11.8.86
objprint==0.3.0
omegaconf==2.3.0
opencv-python==4.11.0.86
opencv-python-headless==4.11.0.86
orjson==3.10.18
packaging==25.0
pandas==2.3.0
paramiko==3.5.1
pillow==11.0.0
platformdirs==4.3.8
plyfile==1.1.2
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pybind11==2.13.6
pycparser==2.22
pydantic==2.11.7
pydantic_core==2.33.2
pydub==0.25.1
pygltflib==1.16.4
Pygments==2.19.2
PyMatting==1.1.14
PyMCubes==0.1.6
pymeshlab==2022.2.post3
PyNaCl==1.5.0
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytorch-lightning==2.5.2
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
regex==2024.11.6
rembg==2.0.66
requests==2.32.4
rich==14.0.0
rosbags==0.10.10
rpds-py==0.25.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
ruff==0.12.1
s3transfer==0.13.0
safehttpx==0.1.6
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.32.0
setproctitle==1.3.6
shellingham==1.5.4
shtab==1.7.2
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
threadpoolctl==3.6.0
tifffile==2025.5.10
tokenizers==0.21.2
tomlkit==0.13.3
torch==2.6.0+cu118
torchaudio==2.6.0+cu118
torchmetrics==1.7.4
torchtyping==0.1.5
torchvision==0.21.0+cu118
tqdm==4.67.1
transformers==4.47.1
trimesh==4.6.13
triton==3.2.0
typeguard==2.13.3
typer==0.16.0
typing-inspect==0.9.0
typing-inspection==0.4.1
typing_extensions==4.14.0
tyro==0.9.25
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.35.0
varname==0.15.0
wandb==0.20.1
websockets==15.0.1
wrapt==1.17.2
xatlas==0.0.10
yarl==1.20.1
zipp==3.23.0
zstandard==0.23.0
